This report, by its very length, defends itself against the risk of being read - code converted to ISO C90 standart

+++ pointers_arrays_strings/2-strlen.c
@@ -8,7 +8,7 @@
  */
 int _strlen(char *s)
 {
-	unsigned long long length;
+	unsigned long length;
 
 	length = 0;
 	while (s[length] != '\0')
